permissions

+++ app/Http/Controllers/DeductController.php
@@ -10,8 +10,15 @@ class DeductController extends Controller
 {
     public function deleteDeduct(Request $request ,Deduct $deduct)
     {
+        $user =  auth()->user();
+
+
+        if ($user->can('حذف اذن طلب')) {
         $deduct->delete();
         return ['success' => true];
+        }else{
+            return \response(['status' => false,'message'=>'صلاحيه  حذف اذن طلب غير مفعله'],400);
+        }
     }
     public function showDeductById(Request $request,Deduct $deduct){
 
@@ -19,19 +26,25 @@ class DeductController extends Controller
     }
     public function getDeductsByDate(Request $request){
         $data = $request->all();
-        $date = Carbon::parse($data['date']);
+        $date = $data['date'];
         $data =  Deduct::WhereDate('created_at','=',$date)->get();
         return ['data'=>$data , 'status'=>true];
     }
     public function destroy(Request $request){
+
         $data = $request->all();
 //        return $data;
         $item_id =  $data['item_id'];
         $deduct = Deduct::latest()->first();
         $deduct->items()->detach($item_id);
         return ['status'=>true];
+
     }
     public function deduct(Request $request){
+        $user =  auth()->user();
+
+
+        if ($user->can('اذن طلب')) {
         $data = $request->all();
 //        return $data;
         $item_id = $data['item_id'];
@@ -47,12 +60,22 @@ class DeductController extends Controller
             'created_At'=>now()
         ]);
         return ['status'=>true];
+        }else{
+            return \response(['status' => false,'message'=>'صلاحيه  اذن طلب  غير مفعله'],400);
+        }
     }
     public function last(){
              return  Deduct::with('items.pivot.client')->latest()->first()    ;
     }
     public function complete(){
+        $user =  auth()->user();
+
+
+        if ($user->can('اذن صرف')) {
         return ['status' => Deduct::create()] ;
+        }else{
+            return \response(['status' => false,'message'=>'صلاحيه  اذن صرف  غير مفعله'],400);
+        }
 
     }
 }
