+++ app/Http/Controllers/SupplierController.php
@@ -8,13 +8,26 @@ use Illuminate\Http\Request;
 class SupplierController extends Controller
 {
     public function destroy(Request $request , Supplier $supplier){
+        $user =  auth()->user();
+
+
+        if ($user->can('حذف مورد')) {
+
+
         return ['status'=>$supplier->delete()];
+        }else{
+            return \response(['status' => false,'msg'=>'صلاحيه حذف مورد غير مفعله'],400);
+
+        }
     }
     public  function index(){
         return Supplier::all();
     }
     public function create(Request $request){
+        $user =  auth()->user();
 
+
+        if ($user->can('اضافه مورد')) {
         $data = $request->all();
         $email = '';
         if (isset($data['email'])){
@@ -29,11 +42,24 @@ class SupplierController extends Controller
 
         ]);
         return ['status'=> $result->id];
+        }else{
+            return \response(['status' => false,'msg'=>'صلاحيه اضافه مورد غير مفعله'],400);
+
+        }
     }
     public function update(Request $request , Supplier $supplier){
-        $data = $request->all();
+        $user =  auth()->user();
+
+
+        if ($user->can('تعديل مورد')) {
+
+            $data = $request->all();
 //        return $data;
 
         return ['status'=>$supplier->update([$data['colName']=>$data['val']])];
+        }else{
+            return \response(['status' => false,'msg'=>'صلاحيه تعديل مورد غير مفعله'],400);
+
+        }
     }
 }
